telegram_bot/requirements.txt
python-telegram-bot==20.5
python-dotenv
supabase==1.1.1